Consolidate Ariakit scripts package helpers
